--- app/models/concerns/module_trim_space.rb
@@ -1,9 +1,0 @@
-module ModuleTrimSpace
-  def trim_space_name
-    self.name = name.squish
-  end
-
-  def trim_space_content
-    self.content = content.squish
-  end
-end
