+++ app/models/concerns/module_trim_space_content.rb
@@ -1,0 +1,13 @@
+module ModuleTrimSpaceContent
+  extend ActiveSupport::Concern
+
+  included do
+    before_save :trim_space_content
+  end
+
+  private
+
+  def trim_space_content
+    self.content = content.squish
+  end
+end
