+++ app/models/user.rb
@@ -1,10 +1,10 @@
 class User < ApplicationRecord
-  include ModuleTrimSpace
+  include ModuleTrimSpaceName
 
   acts_as_paranoid
 
   attr_accessor :remember_token, :activation_token
-  before_save   :downcase_email, :trim_space_name
+  before_save   :downcase_email
   before_create :create_activation_digest
 
   PASSWORD_EXPIRED_TIME = 1
